more robust person thumbnail generation
