+++ server/test/fixtures/person.stub.ts
@@ -178,8 +178,7 @@ export const personThumbnailStub = {
     oldWidth: 2160,
     type: AssetType.IMAGE,
     originalPath: '/original/path.jpg',
-    exifImageHeight: 3840,
-    exifImageWidth: 2160,
+    exifOrientation: '1',
     previewPath: previewFile.path,
   }),
   newThumbnailMiddle: Object.freeze({
@@ -192,8 +191,7 @@ export const personThumbnailStub = {
     oldWidth: 400,
     type: AssetType.IMAGE,
     originalPath: '/original/path.jpg',
-    exifImageHeight: 1000,
-    exifImageWidth: 1000,
+    exifOrientation: '1',
     previewPath: previewFile.path,
   }),
   newThumbnailEnd: Object.freeze({
@@ -206,8 +204,20 @@ export const personThumbnailStub = {
     oldWidth: 500,
     type: AssetType.IMAGE,
     originalPath: '/original/path.jpg',
-    exifImageHeight: 1000,
-    exifImageWidth: 1000,
+    exifOrientation: '1',
+    previewPath: previewFile.path,
+  }),
+  rawEmbeddedThumbnail: Object.freeze({
+    ownerId: userStub.admin.id,
+    x1: 100,
+    y1: 100,
+    x2: 200,
+    y2: 200,
+    oldHeight: 500,
+    oldWidth: 400,
+    type: AssetType.IMAGE,
+    originalPath: '/original/path.dng',
+    exifOrientation: '1',
     previewPath: previewFile.path,
   }),
 };
